Split token listener into a separate file
